refactor(panels): move to new separate folder
