+++ src/pages/ValidatorManagement/ManageValidator/RightManageSettings/Panels/DepositPanel.tsx
